fix: Move `io_uring` into module under `vmm`
