--- src/io_uring/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "io_uring"
-version = "0.1.0"
-authors = ["Amazon Firecracker team <firecracker-devel@amazon.com>"]
-edition = "2021"
-license = "Apache-2.0"
-
-[dependencies]
-derive_more = { version = "0.99.17", default-features = false, features = ["from"] }
-libc = "0.2.117"
-
-utils = { path = "../utils" }
-vm-memory = { path="../vm-memory" }
-
-[dev-dependencies]
-proptest = { version = "1.0.0", default-features = false, features = ["std"] }
