+++ src/vmm/src/io_uring/mod.rs
@@ -1,18 +1,6 @@
 // Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-#![deny(missing_docs)]
-
-//! High-level interface over Linux io_uring.
-//!
-//! Aims to provide an easy-to-use interface, while making some Firecracker-specific simplifying
-//! assumptions. The crate does not currently aim at supporting all io_uring features and use
-//! cases. For example, it only works with pre-registered fds and read/write/fsync requests.
-//!
-//! Requires at least kernel version 5.10.51.
-//! For more information on io_uring, refer to the man pages.
-//! [This pdf](https://kernel.dk/io_uring.pdf) is also very useful, though outdated at times.
-
 #[allow(clippy::undocumented_unsafe_blocks)]
 mod bindings;
 pub mod operation;
